sowm: Shuffle declarations around
